requirements/local.txt
# Local development dependencies go here
-r base.txt
django-debug-toolbar==3.2.1
django-extensions==3.1.3
collection-json==0.1.1
coverage==5.5
pylint==2.9.1
flake8==3.9.2
isort==5.9.1
